+++ .env.example
@@ -1,0 +1,66 @@
+# ==========================================
+# Telemt Easy Configuration
+# ==========================================
+# This file lets you configure your Telegram MTProto proxy easily.
+# If you are using Docker, Railway, or Dokploy, you can copy these variables 
+# directly into your environment settings.
+
+# --- 1. Your Proxy Address (Share Link) ---
+# This is the IP address or domain name you will give to your users so they can connect.
+# Example: If your server IP is 123.45.67.89, put that here.
+# TELEMT_GENERAL__LINKS__PUBLIC_HOST=123.45.67.89
+
+# The port your users will connect to.
+# On Dokploy, the proxy runs on port 8443 by default (port 443 is used by
+# Traefik for HTTP services). Set this to match PROXY_PUBLIC_PORT below.
+# TELEMT_GENERAL__LINKS__PUBLIC_PORT=8443
+
+
+# --- 2. User Passwords (Secrets) ---
+# To connect to your proxy, your users need a "secret" (password).
+# A secret MUST be exactly 32 characters long, using only numbers (0-9) and letters (a-f).
+# Generate a random one with: openssl rand -hex 16
+#
+# Format: TELEMT_ACCESS__USERS__<USERNAME>=<32_CHARACTER_SECRET>
+#
+# Example user "myfamily" with a 32-character secret:
+# TELEMT_ACCESS__USERS__MYFAMILY=11112222333344445555666677778888
+
+
+# --- 3. Anti-Censorship (Fake Domain) ---
+# To hide your proxy from internet censors, telemt disguises the traffic to look like 
+# someone is just browsing a normal website.
+# Enter a popular, unblocked website domain in your country (e.g. google.com, or a local site).
+# TELEMT_CENSORSHIP__TLS_DOMAIN=petrovich.ru
+
+# Keep these enabled for maximum protection against deep packet inspection (DPI).
+# TELEMT_CENSORSHIP__MASK=true
+# TELEMT_CENSORSHIP__TLS_EMULATION=true
+
+
+# --- 4. Sponsored Channel (Optional) ---
+# If you registered your proxy with @MTProxybot on Telegram to show a sponsored channel,
+# paste the 32-character "ad_tag" they gave you here. Leave commented out if you don't have one.
+# TELEMT_GENERAL__AD_TAG=00000000000000000000000000000000
+
+
+# --- 5. Advanced Server Settings ---
+# The port the proxy actually runs on internally. Usually matches your PUBLIC_PORT.
+# TELEMT_SERVER__PORT=443
+
+# How much detail to print in the logs (silent, normal, debug)
+# TELEMT_GENERAL__LOG_LEVEL=normal
+
+
+# --- 6. Dokploy (docker-compose-dokploy.yml only) ---
+# These are used by the Dokploy compose file, NOT Telemt config vars.
+# Do not prefix them with TELEMT_.
+
+# The host port to publish for MTProto traffic. The container always
+# listens on 443 internally; this maps it to a host port that doesn't
+# conflict with Traefik (which owns port 443 for HTTP services).
+# Default is 8443 if not set.
+# PROXY_PUBLIC_PORT=8443
+
+# Log level for the Rust process itself (not Telemt config).
+# RUST_LOG=info
