+++ src/metrics/render.rs
@@ -1,432 +1,19 @@
 use std::collections::{BTreeSet, HashMap};
-use std::convert::Infallible;
-use std::net::SocketAddr;
-use std::sync::Arc;
-use std::time::Duration;
-
-use arc_swap::ArcSwap;
-use http_body_util::Full;
-use hyper::body::Bytes;
-use hyper::server::conn::http1;
-use hyper::service::service_fn;
-use hyper::{Request, Response, StatusCode};
-use tokio::net::TcpListener;
-use tokio::sync::Semaphore;
-use tokio::time::timeout;
-use tracing::{debug, info, warn};
 
 use crate::config::ProxyConfig;
 use crate::ip_tracker::UserIpTracker;
-use crate::maestro::generation::RuntimeGeneration;
 use crate::proxy::shared_state::ProxySharedState;
 use crate::stats::Stats;
-use crate::stats::beobachten::BeobachtenStore;
 use crate::tls_front::TlsFrontCache;
 use crate::tls_front::cache;
 use crate::tls_front::fetcher;
-use crate::transport::{ListenOptions, create_listener};
+
+use super::tls_front::render_tls_front_profile_health;
 
 // Keeps `/metrics` response size bounded when per-user telemetry is enabled.
 const USER_LABELED_METRICS_MAX_USERS: usize = 4096;
-// Keeps TLS-front per-domain health series bounded for large generated configs.
-const TLS_FRONT_PROFILE_HEALTH_MAX_DOMAINS: usize = 256;
-const METRICS_MAX_CONTROL_CONNECTIONS: usize = 512;
-const METRICS_HTTP_CONNECTION_TIMEOUT: Duration = Duration::from_secs(15);
 
-pub async fn serve(
-    port: u16,
-    listen: Option<String>,
-    listen_backlog: u32,
-    active_runtime: Arc<ArcSwap<RuntimeGeneration>>,
-) {
-    // If `metrics_listen` is set, bind on that single address only.
-    if let Some(ref listen_addr) = listen {
-        let addr: SocketAddr = match listen_addr.parse() {
-            Ok(a) => a,
-            Err(e) => {
-                warn!(error = %e, "Invalid metrics_listen address: {}", listen_addr);
-                return;
-            }
-        };
-        // Match `server.api.listen`: `[::]:port` is a dual-stack wildcard
-        // on Linux when `net.ipv6.bindv6only=0`.
-        let ipv6_only = addr.is_ipv6() && !addr.ip().is_unspecified();
-        match bind_metrics_listener(addr, ipv6_only, listen_backlog) {
-            Ok(listener) => {
-                info!("Metrics endpoint: http://{}/metrics and /beobachten", addr);
-                serve_listener(listener, active_runtime).await;
-            }
-            Err(e) => {
-                warn!(error = %e, "Failed to bind metrics on {}", addr);
-            }
-        }
-        return;
-    }
-
-    // Fallback: keep metrics local unless an explicit metrics_listen is configured.
-    let mut listener_v4 = None;
-    let mut listener_v6 = None;
-
-    let addr_v4 = SocketAddr::from(([127, 0, 0, 1], port));
-    match bind_metrics_listener(addr_v4, false, listen_backlog) {
-        Ok(listener) => {
-            info!(
-                "Metrics endpoint: http://{}/metrics and /beobachten",
-                addr_v4
-            );
-            listener_v4 = Some(listener);
-        }
-        Err(e) => {
-            warn!(error = %e, "Failed to bind metrics on {}", addr_v4);
-        }
-    }
-
-    let addr_v6 = SocketAddr::from(([0, 0, 0, 0, 0, 0, 0, 1], port));
-    match bind_metrics_listener(addr_v6, true, listen_backlog) {
-        Ok(listener) => {
-            info!(
-                "Metrics endpoint: http://[::1]:{}/metrics and /beobachten",
-                port
-            );
-            listener_v6 = Some(listener);
-        }
-        Err(e) => {
-            warn!(error = %e, "Failed to bind metrics on {}", addr_v6);
-        }
-    }
-
-    match (listener_v4, listener_v6) {
-        (None, None) => {
-            warn!("Metrics listener is unavailable on both IPv4 and IPv6");
-        }
-        (Some(listener), None) | (None, Some(listener)) => {
-            serve_listener(listener, active_runtime).await;
-        }
-        (Some(listener4), Some(listener6)) => {
-            let active_runtime_v6 = active_runtime.clone();
-            tokio::spawn(async move {
-                serve_listener(listener6, active_runtime_v6).await;
-            });
-            serve_listener(listener4, active_runtime).await;
-        }
-    }
-}
-
-fn bind_metrics_listener(
-    addr: SocketAddr,
-    ipv6_only: bool,
-    listen_backlog: u32,
-) -> std::io::Result<TcpListener> {
-    let options = ListenOptions {
-        reuse_port: false,
-        ipv6_only,
-        backlog: listen_backlog,
-        ..Default::default()
-    };
-    let socket = create_listener(addr, &options)?;
-    TcpListener::from_std(socket.into())
-}
-
-async fn serve_listener(listener: TcpListener, active_runtime: Arc<ArcSwap<RuntimeGeneration>>) {
-    let connection_permits = Arc::new(Semaphore::new(METRICS_MAX_CONTROL_CONNECTIONS));
-
-    loop {
-        let (stream, peer) = match listener.accept().await {
-            Ok(v) => v,
-            Err(e) => {
-                warn!(error = %e, "Metrics accept error");
-                continue;
-            }
-        };
-
-        let runtime = active_runtime.load_full();
-        let config = runtime.config();
-        if !config.server.metrics_whitelist.is_empty()
-            && !config
-                .server
-                .metrics_whitelist
-                .iter()
-                .any(|net| net.contains(peer.ip()))
-        {
-            debug!(peer = %peer, "Metrics request denied by whitelist");
-            continue;
-        }
-
-        let connection_permit = match connection_permits.clone().try_acquire_owned() {
-            Ok(permit) => permit,
-            Err(_) => {
-                debug!(
-                    peer = %peer,
-                    max_connections = METRICS_MAX_CONTROL_CONNECTIONS,
-                    "Dropping metrics connection: control-plane connection budget exhausted"
-                );
-                continue;
-            }
-        };
-
-        let active_runtime = active_runtime.clone();
-        tokio::spawn(async move {
-            let _connection_permit = connection_permit;
-            let svc = service_fn(move |req| {
-                let runtime = active_runtime.load_full();
-                let stats = runtime.stats.clone();
-                let beobachten = runtime.beobachten.clone();
-                let shared_state = runtime.proxy_shared.clone();
-                let ip_tracker = runtime.ip_tracker.clone();
-                let tls_cache = runtime.tls_cache.clone();
-                let config = runtime.config();
-                async move {
-                    handle(
-                        req,
-                        &stats,
-                        &beobachten,
-                        &shared_state,
-                        &ip_tracker,
-                        tls_cache.as_deref(),
-                        &config,
-                    )
-                    .await
-                }
-            });
-            match timeout(
-                METRICS_HTTP_CONNECTION_TIMEOUT,
-                http1::Builder::new().serve_connection(hyper_util::rt::TokioIo::new(stream), svc),
-            )
-            .await
-            {
-                Ok(Ok(())) => {}
-                Ok(Err(e)) => {
-                    debug!(error = %e, "Metrics connection error");
-                }
-                Err(_) => {
-                    debug!(
-                        peer = %peer,
-                        timeout_ms = METRICS_HTTP_CONNECTION_TIMEOUT.as_millis() as u64,
-                        "Metrics connection timed out"
-                    );
-                }
-            }
-        });
-    }
-}
-
-async fn handle<B>(
-    req: Request<B>,
-    stats: &Stats,
-    beobachten: &BeobachtenStore,
-    shared_state: &ProxySharedState,
-    ip_tracker: &UserIpTracker,
-    tls_cache: Option<&TlsFrontCache>,
-    config: &ProxyConfig,
-) -> Result<Response<Full<Bytes>>, Infallible> {
-    if req.uri().path() == "/metrics" {
-        let body = render_metrics(stats, shared_state, config, ip_tracker, tls_cache).await;
-        let resp = Response::builder()
-            .status(StatusCode::OK)
-            .header("content-type", "text/plain; version=0.0.4; charset=utf-8")
-            .body(Full::new(Bytes::from(body)))
-            .unwrap();
-        return Ok(resp);
-    }
-
-    if req.uri().path() == "/beobachten" {
-        let body = render_beobachten(stats, beobachten, config);
-        let resp = Response::builder()
-            .status(StatusCode::OK)
-            .header("content-type", "text/plain; charset=utf-8")
-            .body(Full::new(Bytes::from(body)))
-            .unwrap();
-        return Ok(resp);
-    }
-
-    let resp = Response::builder()
-        .status(StatusCode::NOT_FOUND)
-        .body(Full::new(Bytes::from("Not Found\n")))
-        .unwrap();
-    Ok(resp)
-}
-
-fn render_beobachten(stats: &Stats, beobachten: &BeobachtenStore, config: &ProxyConfig) -> String {
-    if !config.general.beobachten {
-        return "beobachten disabled\n".to_string();
-    }
-
-    let ttl = Duration::from_secs(config.general.beobachten_minutes.saturating_mul(60));
-    let mut body = beobachten.snapshot_text(ttl);
-    let tls_text = stats.tls_fingerprint_snapshot_text(ttl, 20);
-    if !tls_text.is_empty() {
-        if !body.ends_with('\n') {
-            body.push('\n');
-        }
-        body.push('\n');
-        body.push_str(&tls_text);
-    }
-    body
-}
-
-fn tls_front_domains(config: &ProxyConfig) -> Vec<String> {
-    let mut domains = Vec::with_capacity(1 + config.censorship.tls_domains.len());
-    if !config.censorship.tls_domain.is_empty() {
-        domains.push(config.censorship.tls_domain.clone());
-    }
-    for domain in &config.censorship.tls_domains {
-        if !domain.is_empty() && !domains.contains(domain) {
-            domains.push(domain.clone());
-        }
-    }
-    domains
-}
-
-fn prometheus_label_value(value: &str) -> String {
-    value.replace('\\', "\\\\").replace('"', "\\\"")
-}
-
-async fn render_tls_front_profile_health(
-    out: &mut String,
-    config: &ProxyConfig,
-    tls_cache: Option<&TlsFrontCache>,
-) {
-    use std::fmt::Write;
-
-    let domains = tls_front_domains(config);
-    let (health, suppressed) = match (config.censorship.tls_emulation, tls_cache) {
-        (true, Some(cache)) => {
-            cache
-                .profile_health_snapshot(&domains, TLS_FRONT_PROFILE_HEALTH_MAX_DOMAINS)
-                .await
-        }
-        _ => (Vec::new(), domains.len()),
-    };
-
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_domains TLS front configured profile domains by export status"
-    );
-    let _ = writeln!(out, "# TYPE telemt_tls_front_profile_domains gauge");
-    let _ = writeln!(
-        out,
-        "telemt_tls_front_profile_domains{{status=\"configured\"}} {}",
-        domains.len()
-    );
-    let _ = writeln!(
-        out,
-        "telemt_tls_front_profile_domains{{status=\"emitted\"}} {}",
-        health.len()
-    );
-    let _ = writeln!(
-        out,
-        "telemt_tls_front_profile_domains{{status=\"suppressed\"}} {}",
-        suppressed
-    );
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_info TLS front profile source and feature flags per configured domain"
-    );
-    let _ = writeln!(out, "# TYPE telemt_tls_front_profile_info gauge");
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_quality_info TLS front profile quality and key-share group per configured domain"
-    );
-    let _ = writeln!(out, "# TYPE telemt_tls_front_profile_quality_info gauge");
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_age_seconds Age of cached TLS front profile data per configured domain"
-    );
-    let _ = writeln!(out, "# TYPE telemt_tls_front_profile_age_seconds gauge");
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_server_hello_bytes TLS front cached ServerHello record body bytes per configured domain"
-    );
-    let _ = writeln!(
-        out,
-        "# TYPE telemt_tls_front_profile_server_hello_bytes gauge"
-    );
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_server_hello_extensions TLS front cached visible ServerHello extension count per configured domain"
-    );
-    let _ = writeln!(
-        out,
-        "# TYPE telemt_tls_front_profile_server_hello_extensions gauge"
-    );
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_app_data_records TLS front cached app-data record count per configured domain"
-    );
-    let _ = writeln!(
-        out,
-        "# TYPE telemt_tls_front_profile_app_data_records gauge"
-    );
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_ticket_records TLS front cached ticket-like tail record count per configured domain"
-    );
-    let _ = writeln!(out, "# TYPE telemt_tls_front_profile_ticket_records gauge");
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_change_cipher_spec_records TLS front cached ChangeCipherSpec record count per configured domain"
-    );
-    let _ = writeln!(
-        out,
-        "# TYPE telemt_tls_front_profile_change_cipher_spec_records gauge"
-    );
-    let _ = writeln!(
-        out,
-        "# HELP telemt_tls_front_profile_app_data_bytes TLS front cached total app-data bytes per configured domain"
-    );
-    let _ = writeln!(out, "# TYPE telemt_tls_front_profile_app_data_bytes gauge");
-
-    for item in health {
-        let domain = prometheus_label_value(&item.domain);
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_info{{domain=\"{}\",source=\"{}\",is_default=\"{}\",has_cert_info=\"{}\",has_cert_payload=\"{}\"}} 1",
-            domain, item.source, item.is_default, item.has_cert_info, item.has_cert_payload
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_quality_info{{domain=\"{}\",quality=\"{}\",key_share_group=\"{}\"}} 1",
-            domain, item.quality, item.key_share_group
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_age_seconds{{domain=\"{}\"}} {}",
-            domain, item.age_seconds
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_server_hello_bytes{{domain=\"{}\"}} {}",
-            domain, item.server_hello_record_len
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_server_hello_extensions{{domain=\"{}\"}} {}",
-            domain, item.server_hello_extensions
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_app_data_records{{domain=\"{}\"}} {}",
-            domain, item.app_data_records
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_ticket_records{{domain=\"{}\"}} {}",
-            domain, item.ticket_records
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_change_cipher_spec_records{{domain=\"{}\"}} {}",
-            domain, item.change_cipher_spec_count
-        );
-        let _ = writeln!(
-            out,
-            "telemt_tls_front_profile_app_data_bytes{{domain=\"{}\"}} {}",
-            domain, item.total_app_data_len
-        );
-    }
-}
-
-async fn render_metrics(
+pub(super) async fn render_metrics(
     stats: &Stats,
     shared_state: &ProxySharedState,
     config: &ProxyConfig,
@@ -3857,437 +3444,4 @@ async fn render_metrics(
     );
 
     out
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use http_body_util::BodyExt;
-    use std::net::IpAddr;
-    use std::time::SystemTime;
-
-    use crate::tls_front::types::{
-        CachedTlsData, ParsedServerHello, TlsBehaviorProfile, TlsCertPayload, TlsProfileSource,
-    };
-
-    #[tokio::test]
-    async fn test_render_metrics_format() {
-        let stats = Arc::new(Stats::new());
-        let shared_state = ProxySharedState::new();
-        let tracker = UserIpTracker::new();
-        let mut config = ProxyConfig::default();
-        config
-            .access
-            .user_max_unique_ips
-            .insert("alice".to_string(), 4);
-
-        stats.increment_connects_all();
-        stats.increment_connects_all();
-        stats.increment_connects_bad_with_class("tls_handshake_bad_client");
-        stats.increment_handshake_timeouts();
-        stats.increment_handshake_failure_class("timeout");
-        shared_state
-            .handshake
-            .auth_expensive_checks_total
-            .fetch_add(9, std::sync::atomic::Ordering::Relaxed);
-        shared_state
-            .handshake
-            .auth_budget_exhausted_total
-            .fetch_add(2, std::sync::atomic::Ordering::Relaxed);
-        stats.increment_upstream_connect_attempt_total();
-        stats.increment_upstream_connect_attempt_total();
-        stats.increment_upstream_connect_success_total();
-        stats.increment_upstream_connect_fail_total();
-        stats.increment_upstream_connect_failfast_hard_error_total();
-        stats.observe_upstream_connect_attempts_per_request(2);
-        stats.observe_upstream_connect_duration_ms(220, true);
-        stats.observe_upstream_connect_duration_ms(1500, false);
-        stats.increment_me_rpc_proxy_req_signal_sent_total();
-        stats.increment_me_rpc_proxy_req_signal_failed_total();
-        stats.increment_me_rpc_proxy_req_signal_skipped_no_meta_total();
-        stats.increment_me_rpc_proxy_req_signal_response_total();
-        stats.increment_me_rpc_proxy_req_signal_close_sent_total();
-        stats.increment_me_idle_close_by_peer_total();
-        stats.increment_relay_idle_soft_mark_total();
-        stats.increment_relay_idle_hard_close_total();
-        stats.increment_relay_pressure_evict_total();
-        stats.increment_relay_protocol_desync_close_total();
-        stats.increment_me_d2c_batches_total();
-        stats.add_me_d2c_batch_frames_total(3);
-        stats.add_me_d2c_batch_bytes_total(2048);
-        stats.increment_me_d2c_flush_reason(crate::stats::MeD2cFlushReason::AckImmediate);
-        stats.increment_me_d2c_data_frames_total();
-        stats.increment_me_d2c_ack_frames_total();
-        stats.add_me_d2c_payload_bytes_total(1800);
-        stats.increment_me_d2c_write_mode(crate::stats::MeD2cWriteMode::Coalesced);
-        stats.increment_me_d2c_quota_reject_total(crate::stats::MeD2cQuotaRejectStage::PostWrite);
-        stats.observe_me_d2c_frame_buf_shrink(4096);
-        stats.increment_me_endpoint_quarantine_total();
-        stats.increment_me_endpoint_quarantine_unexpected_total();
-        stats.increment_me_endpoint_quarantine_draining_suppressed_total();
-        stats.increment_user_connects("alice");
-        stats.increment_user_curr_connects("alice");
-        stats.add_user_octets_from("alice", 1024);
-        stats.add_user_octets_to("alice", 2048);
-        stats.increment_user_msgs_from("alice");
-        stats.increment_user_msgs_to("alice");
-        stats.increment_user_msgs_to("alice");
-        tracker
-            .check_and_add("alice", "203.0.113.10".parse().unwrap())
-            .await
-            .unwrap();
-
-        let output = render_metrics(&stats, shared_state.as_ref(), &config, &tracker, None).await;
-
-        assert!(output.contains(&format!(
-            "telemt_build_info{{version=\"{}\"}} 1",
-            env!("CARGO_PKG_VERSION")
-        )));
-        assert!(output.contains("telemt_connections_total 2"));
-        assert!(output.contains("telemt_connections_bad_total 1"));
-        assert!(output.contains(
-            "telemt_connections_bad_by_class_total{class=\"tls_handshake_bad_client\"} 1"
-        ));
-        assert!(output.contains("telemt_handshake_timeouts_total 1"));
-        assert!(output.contains("telemt_handshake_failures_by_class_total{class=\"timeout\"} 1"));
-        assert!(output.contains("telemt_auth_expensive_checks_total 9"));
-        assert!(output.contains("telemt_auth_budget_exhausted_total 2"));
-        assert!(output.contains("telemt_upstream_connect_attempt_total 2"));
-        assert!(output.contains("telemt_upstream_connect_success_total 1"));
-        assert!(output.contains("telemt_upstream_connect_fail_total 1"));
-        assert!(output.contains("telemt_upstream_connect_failfast_hard_error_total 1"));
-        assert!(output.contains("telemt_upstream_connect_attempts_per_request{bucket=\"2\"} 1"));
-        assert!(
-            output
-                .contains("telemt_upstream_connect_duration_success_total{bucket=\"101_500ms\"} 1")
-        );
-        assert!(
-            output.contains("telemt_upstream_connect_duration_fail_total{bucket=\"gt_1000ms\"} 1")
-        );
-        assert!(output.contains("telemt_me_rpc_proxy_req_signal_sent_total 1"));
-        assert!(output.contains("telemt_me_rpc_proxy_req_signal_failed_total 1"));
-        assert!(output.contains("telemt_me_rpc_proxy_req_signal_skipped_no_meta_total 1"));
-        assert!(output.contains("telemt_me_rpc_proxy_req_signal_response_total 1"));
-        assert!(output.contains("telemt_me_rpc_proxy_req_signal_close_sent_total 1"));
-        assert!(output.contains("telemt_me_idle_close_by_peer_total 1"));
-        assert!(output.contains("telemt_relay_idle_soft_mark_total 1"));
-        assert!(output.contains("telemt_relay_idle_hard_close_total 1"));
-        assert!(output.contains("telemt_relay_pressure_evict_total 1"));
-        assert!(output.contains("telemt_relay_protocol_desync_close_total 1"));
-        assert!(output.contains("telemt_me_d2c_batches_total 1"));
-        assert!(output.contains("telemt_me_d2c_batch_frames_total 3"));
-        assert!(output.contains("telemt_me_d2c_batch_bytes_total 2048"));
-        assert!(output.contains("telemt_me_d2c_flush_reason_total{reason=\"ack_immediate\"} 1"));
-        assert!(output.contains("telemt_me_d2c_data_frames_total 1"));
-        assert!(output.contains("telemt_me_d2c_ack_frames_total 1"));
-        assert!(output.contains("telemt_me_d2c_payload_bytes_total 1800"));
-        assert!(output.contains("telemt_me_d2c_write_mode_total{mode=\"coalesced\"} 1"));
-        assert!(output.contains("telemt_me_d2c_quota_reject_total{stage=\"post_write\"} 1"));
-        assert!(output.contains("telemt_me_d2c_frame_buf_shrink_total 1"));
-        assert!(output.contains("telemt_me_d2c_frame_buf_shrink_bytes_total 4096"));
-        assert!(output.contains("telemt_me_endpoint_quarantine_total 1"));
-        assert!(output.contains("telemt_me_endpoint_quarantine_unexpected_total 1"));
-        assert!(output.contains("telemt_me_endpoint_quarantine_draining_suppressed_total 1"));
-        assert!(output.contains("telemt_user_connections_total{user=\"alice\"} 1"));
-        assert!(output.contains("telemt_user_connections_current{user=\"alice\"} 1"));
-        assert!(output.contains("telemt_user_octets_from_client{user=\"alice\"} 1024"));
-        assert!(output.contains("telemt_user_octets_to_client{user=\"alice\"} 2048"));
-        assert!(output.contains("telemt_user_msgs_from_client{user=\"alice\"} 1"));
-        assert!(output.contains("telemt_user_msgs_to_client{user=\"alice\"} 2"));
-        assert!(output.contains("telemt_user_unique_ips_current{user=\"alice\"} 1"));
-        assert!(output.contains("telemt_user_unique_ips_recent_window{user=\"alice\"} 1"));
-        assert!(output.contains("telemt_user_unique_ips_limit{user=\"alice\"} 4"));
-        assert!(output.contains("telemt_user_unique_ips_utilization{user=\"alice\"} 0.250000"));
-        assert!(output.contains("telemt_ip_tracker_users{scope=\"active\"} 1"));
-        assert!(output.contains("telemt_ip_tracker_entries{scope=\"active\"} 1"));
-        assert!(output.contains("telemt_ip_tracker_cleanup_queue_len 0"));
-    }
-
-    #[tokio::test]
-    async fn test_render_tls_front_profile_health() {
-        let stats = Stats::new();
-        let shared_state = ProxySharedState::new();
-        let tracker = UserIpTracker::new();
-        let mut config = ProxyConfig::default();
-        config.censorship.tls_domain = "primary.example".to_string();
-        config.censorship.tls_domains = vec!["fallback.example".to_string()];
-
-        let cache = TlsFrontCache::new(
-            &[
-                "primary.example".to_string(),
-                "fallback.example".to_string(),
-            ],
-            1024,
-            "tlsfront-profile-health-test",
-        );
-        cache
-            .set(
-                "primary.example",
-                CachedTlsData {
-                    server_hello_template: ParsedServerHello {
-                        version: [0x03, 0x03],
-                        random: [0u8; 32],
-                        session_id: Vec::new(),
-                        cipher_suite: [0x13, 0x01],
-                        compression: 0,
-                        extensions: {
-                            let mut key_share = vec![0x00, 0x1d, 0x00, 0x20];
-                            key_share.resize(36, 0x42);
-                            vec![
-                                crate::tls_front::types::TlsExtension {
-                                    ext_type: 0x002b,
-                                    data: vec![0x03, 0x04],
-                                },
-                                crate::tls_front::types::TlsExtension {
-                                    ext_type: 0x0033,
-                                    data: key_share,
-                                },
-                            ]
-                        },
-                    },
-                    cert_info: None,
-                    cert_payload: Some(TlsCertPayload {
-                        cert_chain_der: vec![vec![0x30, 0x01]],
-                        certificate_message: vec![0x0b, 0x00, 0x00, 0x00],
-                    }),
-                    app_data_records_sizes: vec![1024, 512],
-                    total_app_data_len: 1536,
-                    behavior_profile: TlsBehaviorProfile {
-                        change_cipher_spec_count: 1,
-                        app_data_record_sizes: vec![1024, 512],
-                        ticket_record_sizes: vec![69],
-                        source: TlsProfileSource::Merged,
-                        ..TlsBehaviorProfile::default()
-                    },
-                    fetched_at: SystemTime::now(),
-                    domain: "primary.example".to_string(),
-                },
-            )
-            .await;
-
-        let output = render_metrics(&stats, &shared_state, &config, &tracker, Some(&cache)).await;
-
-        assert!(output.contains("telemt_tls_front_profile_domains{status=\"configured\"} 2"));
-        assert!(output.contains("telemt_tls_front_profile_domains{status=\"emitted\"} 2"));
-        assert!(output.contains("telemt_tls_front_profile_domains{status=\"suppressed\"} 0"));
-        assert!(
-            output.contains("telemt_tls_front_profile_info{domain=\"primary.example\",source=\"merged\",is_default=\"false\",has_cert_info=\"false\",has_cert_payload=\"true\"} 1")
-        );
-        assert!(
-            output.contains("telemt_tls_front_profile_info{domain=\"fallback.example\",source=\"default\",is_default=\"true\",has_cert_info=\"false\",has_cert_payload=\"false\"} 1")
-        );
-        assert!(
-            output.contains("telemt_tls_front_profile_quality_info{domain=\"primary.example\",quality=\"raw_strict\",key_share_group=\"x25519\"} 1")
-        );
-        assert!(
-            output.contains("telemt_tls_front_profile_quality_info{domain=\"fallback.example\",quality=\"fallback\",key_share_group=\"none\"} 1")
-        );
-        assert!(output.contains(
-            "telemt_tls_front_profile_server_hello_bytes{domain=\"primary.example\"} 90"
-        ));
-        assert!(output.contains(
-            "telemt_tls_front_profile_server_hello_extensions{domain=\"primary.example\"} 2"
-        ));
-        assert!(
-            output.contains(
-                "telemt_tls_front_profile_app_data_records{domain=\"primary.example\"} 2"
-            )
-        );
-        assert!(
-            output
-                .contains("telemt_tls_front_profile_ticket_records{domain=\"primary.example\"} 1")
-        );
-        assert!(output.contains(
-            "telemt_tls_front_profile_change_cipher_spec_records{domain=\"primary.example\"} 1"
-        ));
-        assert!(
-            output.contains(
-                "telemt_tls_front_profile_app_data_bytes{domain=\"primary.example\"} 1536"
-            )
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_empty_stats() {
-        let stats = Stats::new();
-        let shared_state = ProxySharedState::new();
-        let tracker = UserIpTracker::new();
-        let config = ProxyConfig::default();
-        let output = render_metrics(&stats, &shared_state, &config, &tracker, None).await;
-        assert!(output.contains("telemt_connections_total 0"));
-        assert!(output.contains("telemt_connections_bad_total 0"));
-        assert!(output.contains("telemt_handshake_timeouts_total 0"));
-        assert!(output.contains("telemt_auth_expensive_checks_total 0"));
-        assert!(output.contains("telemt_auth_budget_exhausted_total 0"));
-        assert!(output.contains("telemt_user_unique_ips_current{user="));
-        assert!(output.contains("telemt_user_unique_ips_recent_window{user="));
-    }
-
-    #[tokio::test]
-    async fn test_render_uses_global_each_unique_ip_limit() {
-        let stats = Stats::new();
-        let shared_state = ProxySharedState::new();
-        stats.increment_user_connects("alice");
-        stats.increment_user_curr_connects("alice");
-        let tracker = UserIpTracker::new();
-        tracker
-            .check_and_add("alice", "203.0.113.10".parse().unwrap())
-            .await
-            .unwrap();
-        let mut config = ProxyConfig::default();
-        config.access.user_max_unique_ips_global_each = 2;
-
-        let output = render_metrics(&stats, &shared_state, &config, &tracker, None).await;
-
-        assert!(output.contains("telemt_user_unique_ips_limit{user=\"alice\"} 2"));
-        assert!(output.contains("telemt_user_unique_ips_utilization{user=\"alice\"} 0.500000"));
-    }
-
-    #[tokio::test]
-    async fn test_render_has_type_annotations() {
-        let stats = Stats::new();
-        let shared_state = ProxySharedState::new();
-        let tracker = UserIpTracker::new();
-        let config = ProxyConfig::default();
-        let output = render_metrics(&stats, &shared_state, &config, &tracker, None).await;
-        assert!(output.contains("# TYPE telemt_uptime_seconds gauge"));
-        assert!(output.contains("# TYPE telemt_connections_total counter"));
-        assert!(output.contains("# TYPE telemt_connections_bad_total counter"));
-        assert!(output.contains("# TYPE telemt_connections_bad_by_class_total counter"));
-        assert!(output.contains("# TYPE telemt_handshake_timeouts_total counter"));
-        assert!(output.contains("# TYPE telemt_handshake_failures_by_class_total counter"));
-        assert!(output.contains("# TYPE telemt_auth_expensive_checks_total counter"));
-        assert!(output.contains("# TYPE telemt_auth_budget_exhausted_total counter"));
-        assert!(output.contains("# TYPE telemt_upstream_connect_attempt_total counter"));
-        assert!(output.contains("# TYPE telemt_me_rpc_proxy_req_signal_sent_total counter"));
-        assert!(output.contains("# TYPE telemt_me_idle_close_by_peer_total counter"));
-        assert!(output.contains("# TYPE telemt_relay_idle_soft_mark_total counter"));
-        assert!(output.contains("# TYPE telemt_relay_idle_hard_close_total counter"));
-        assert!(output.contains("# TYPE telemt_relay_pressure_evict_total counter"));
-        assert!(output.contains("# TYPE telemt_relay_protocol_desync_close_total counter"));
-        assert!(output.contains("# TYPE telemt_me_d2c_batches_total counter"));
-        assert!(output.contains("# TYPE telemt_me_d2c_flush_reason_total counter"));
-        assert!(output.contains("# TYPE telemt_me_d2c_write_mode_total counter"));
-        assert!(output.contains("# TYPE telemt_me_d2c_batch_frames_bucket_total counter"));
-        assert!(output.contains("# TYPE telemt_me_d2c_flush_duration_us_bucket_total counter"));
-        assert!(output.contains("# TYPE telemt_me_endpoint_quarantine_total counter"));
-        assert!(output.contains("# TYPE telemt_me_endpoint_quarantine_unexpected_total counter"));
-        assert!(
-            output
-                .contains("# TYPE telemt_me_endpoint_quarantine_draining_suppressed_total counter")
-        );
-        assert!(output.contains("# TYPE telemt_me_writer_removed_total counter"));
-        assert!(
-            output
-                .contains("# TYPE telemt_me_writer_removed_unexpected_minus_restored_total gauge")
-        );
-        assert!(output.contains("# TYPE telemt_user_unique_ips_current gauge"));
-        assert!(output.contains("# TYPE telemt_user_unique_ips_recent_window gauge"));
-        assert!(output.contains("# TYPE telemt_user_unique_ips_limit gauge"));
-        assert!(output.contains("# TYPE telemt_user_unique_ips_utilization gauge"));
-        assert!(output.contains("# TYPE telemt_stats_user_entries gauge"));
-        assert!(output.contains("# TYPE telemt_telemetry_user_series_users gauge"));
-        assert!(output.contains("# TYPE telemt_ip_tracker_users gauge"));
-        assert!(output.contains("# TYPE telemt_ip_tracker_entries gauge"));
-        assert!(output.contains("# TYPE telemt_ip_tracker_cleanup_queue_len gauge"));
-        assert!(output.contains("# TYPE telemt_ip_tracker_cleanup_total counter"));
-        assert!(output.contains("# TYPE telemt_ip_tracker_cap_rejects_total counter"));
-        assert!(output.contains("# TYPE telemt_tls_fetch_profile_cache_entries gauge"));
-        assert!(output.contains("# TYPE telemt_tls_fetch_profile_cache_cap_drops_total counter"));
-        assert!(output.contains("# TYPE telemt_tls_front_full_cert_budget_ips gauge"));
-        assert!(
-            output.contains("# TYPE telemt_tls_front_full_cert_budget_cap_drops_total counter")
-        );
-        assert!(output.contains("# TYPE telemt_tls_front_profile_domains gauge"));
-        assert!(output.contains("# TYPE telemt_tls_front_profile_info gauge"));
-        assert!(output.contains("# TYPE telemt_tls_front_profile_quality_info gauge"));
-        assert!(output.contains("# TYPE telemt_tls_front_profile_age_seconds gauge"));
-        assert!(output.contains("# TYPE telemt_tls_front_profile_server_hello_bytes gauge"));
-        assert!(output.contains("# TYPE telemt_tls_front_profile_server_hello_extensions gauge"));
-        assert!(output.contains("# TYPE telemt_tls_front_profile_app_data_records gauge"));
-        assert!(output.contains("# TYPE telemt_tls_front_profile_ticket_records gauge"));
-        assert!(
-            output.contains("# TYPE telemt_tls_front_profile_change_cipher_spec_records gauge")
-        );
-        assert!(output.contains("# TYPE telemt_tls_front_profile_app_data_bytes gauge"));
-    }
-
-    #[tokio::test]
-    async fn test_endpoint_integration() {
-        let stats = Arc::new(Stats::new());
-        let beobachten = Arc::new(BeobachtenStore::new());
-        let shared_state = ProxySharedState::new();
-        let tracker = UserIpTracker::new();
-        let mut config = ProxyConfig::default();
-        stats.increment_connects_all();
-        stats.increment_connects_all();
-        stats.increment_connects_all();
-
-        let req = Request::builder().uri("/metrics").body(()).unwrap();
-        let resp = handle(
-            req,
-            &stats,
-            &beobachten,
-            shared_state.as_ref(),
-            &tracker,
-            None,
-            &config,
-        )
-        .await
-        .unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        let body = resp.into_body().collect().await.unwrap().to_bytes();
-        assert!(
-            std::str::from_utf8(body.as_ref())
-                .unwrap()
-                .contains("telemt_connections_total 3")
-        );
-        assert!(
-            std::str::from_utf8(body.as_ref())
-                .unwrap()
-                .contains(&format!(
-                    "telemt_build_info{{version=\"{}\"}} 1",
-                    env!("CARGO_PKG_VERSION")
-                ))
-        );
-
-        config.general.beobachten = true;
-        config.general.beobachten_minutes = 10;
-        beobachten.record(
-            "TLS-scanner",
-            "203.0.113.10".parse::<IpAddr>().unwrap(),
-            Duration::from_secs(600),
-        );
-        let req_beob = Request::builder().uri("/beobachten").body(()).unwrap();
-        let resp_beob = handle(
-            req_beob,
-            &stats,
-            &beobachten,
-            shared_state.as_ref(),
-            &tracker,
-            None,
-            &config,
-        )
-        .await
-        .unwrap();
-        assert_eq!(resp_beob.status(), StatusCode::OK);
-        let body_beob = resp_beob.into_body().collect().await.unwrap().to_bytes();
-        let beob_text = std::str::from_utf8(body_beob.as_ref()).unwrap();
-        assert!(beob_text.contains("[TLS-scanner]"));
-        assert!(beob_text.contains("203.0.113.10-1"));
-
-        let req404 = Request::builder().uri("/other").body(()).unwrap();
-        let resp404 = handle(
-            req404,
-            &stats,
-            &beobachten,
-            shared_state.as_ref(),
-            &tracker,
-            None,
-            &config,
-        )
-        .await
-        .unwrap();
-        assert_eq!(resp404.status(), StatusCode::NOT_FOUND);
-    }
 }
